# backend/scripts/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DataLabellingSupportSystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/exports/yolo/tasks/{taskId}": {
      "get": {
        "tags": [
          "Exports"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "LabelingTasks"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "LabelingTasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/submitted": {
      "get": {
        "tags": [
          "LabelingTasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/manager/labels/{labelId}": {
      "put": {
        "tags": [
          "Manager"
        ],
        "parameters": [
          {
            "name": "labelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Manager.UpdateLabelRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Manager.UpdateLabelRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Manager.UpdateLabelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Manager"
        ],
        "parameters": [
          {
            "name": "labelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/review_errors/review/{id}": {
      "get": {
        "tags": [
          "ReviewErrors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.AddReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.AddReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.AddReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Manager.UpdateLabelRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "yoloClassId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "annotationTypeId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.AddReviewRequest": {
        "type": "object",
        "properties": {
          "annotationSetId": {
            "type": "string",
            "nullable": true
          },
          "reviewerId": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateReviewRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloExportResponse": {
        "type": "object",
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloLabelFileResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Responses.Exports.YoloLabelFileResponse": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Manager.LabelResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "yoloClassId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "annotationTypeId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Models.DataItem": {
        "required": [
          "datasetId",
          "id",
          "objectKey",
          "storageProvider"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "datasetId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "dataset": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.Dataset"
          },
          "storageProvider": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          },
          "objectKey": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "originalWidth": {
            "type": "integer",
            "format": "int32"
          },
          "originalHeight": {
            "type": "integer",
            "format": "int32"
          },
          "dataType": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "checksum": {
            "maxLength": 64,
            "type": "string",
            "nullable": true
          },
          "status": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "uploadedByUserId": {
            "maxLength": 24,
            "type": "string",
            "nullable": true
          },
          "uploadedByUser": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Models.Dataset": {
        "required": [
          "id",
          "name",
          "projectId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "projectId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.Project"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Models.LabelingTask": {
        "required": [
          "annotatorId",
          "dataItemId",
          "id",
          "projectId",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "projectId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.Project"
          },
          "dataItemId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "dataItem": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.DataItem"
          },
          "annotatorId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "annotator": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.User"
          },
          "assignedByUserId": {
            "maxLength": 24,
            "type": "string",
            "nullable": true
          },
          "assignedByUser": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.User"
          },
          "status": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Models.Project": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "guideline": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Models.Role": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Models.User": {
        "required": [
          "email",
          "fullName",
          "id",
          "passwordHash",
          "roleId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "fullName": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "identifyNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "gender": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 300,
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "passwordHash": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "roleId": {
            "maxLength": 24,
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.Role"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste only the JWT access token here. Swagger will send: Authorization: Bearer <token>",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}